Softwarestand nach Abschluss der Vorlesung

+++ src/clubadmin/PMapTest.java
@@ -11,10 +11,9 @@ public class PMapTest {
 		Member m1 = new ActiveMember("Otto", 37);
 		Member m2 = new ActiveMember(m1.createPMap());
 		assertEquals("Otto", m2.getName());
-		assertEquals(37, m2.getAge());
+		assertEquals(37,m2.getAge().intValue());
 		assertEquals(m1.getType(), m2.getType());
-		assertEquals("active", m2.getType());
-		
+		assertEquals("active", m2.getType());		
 	}
 
 }
